design: 현재 시간으로 보여주게 수정
